* Moved data structs into separate file.
